requirements.txt
audioread==3.0.1
Authlib==1.2.1
blinker==1.7.0
boto3==1.29.2
botocore==1.32.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
cryptography==41.0.5
decorator==5.1.1
dnspython==2.4.2
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
future==0.18.3
gevent==23.9.1
greenlet==3.0.1
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
loguru==0.7.2
MarkupSafe==2.1.3
msgpack==1.0.7
numba==0.58.1
numpy==1.26.2
packaging==23.2
pinecone-client==2.2.4
platformdirs==4.0.0
pooch==1.8.0
pycparser==2.21
pymongo==4.6.0
PySocks==1.7.1
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
requests-file==1.5.1
s3transfer==0.7.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
threadpoolctl==3.2.0
tldextract==5.1.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
validators==0.22.0
weaviate-client==3.25.3
Werkzeug==3.0.1
zope.event==5.0
zope.interface==6.1
